backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1
pydantic==2.7.1
pdfplumber==0.11.0
pytesseract==0.3.10
pdf2image==1.17.0
openai==1.30.1
python-dateutil==2.9.0.post0
